perf(cli): isolate leaderboard calculation
