+++ apps/api/src/providers/score.ts
@@ -1,0 +1,8 @@
+import { config } from '@rctf/config'
+import { scoreProviders, type ScoreProvider } from '@rctf/scoring'
+import { loadProvider } from './load'
+
+export const scoreProvider = loadProvider<ScoreProvider>(
+  scoreProviders,
+  config.scoreProvider
+)!
