+++ apps/api/src/services/solve-points.ts
@@ -11,7 +11,7 @@ import { and, asc, eq, sql } from 'drizzle-orm'
 import type { PinoLogger } from 'hono-pino'
 import type { TypedRedis } from '../cache/scripts'
 import { inJsonbArray, insertInChunks } from '../lib/db-bulk'
-import { scoreProvider } from '../providers'
+import { scoreProvider } from '../providers/score'
 import { requestChallengeRecompute } from '../workers'
 import {
   getDecayChallenge,
